Accelerate aarch64 implementation
Take advantage of the vpaddlq_s16 instruction to avoid costly calls to
vuzp{1,2}q_s16.

The overflow behavior is slightly different from the previous
implementation (and from the Intel implementation): intermediate
horizontal adds between 16 bits accumulators are not saturated and the
registers are widen instead, which gives more accuracy, so it should not
be an issue.